refactor: complete file hierarchy refactor
