+++ browser/index.js
@@ -8,12 +8,12 @@ import ReactDOM from 'react-dom'
 import Simulation from './simulation'
 import SimulationContainer from './container/Simulation'
 import Menu from './menu/menu'
-import Navigator from './navigator'
-import Intro from './intro.jsx'
-import introText from './introText.js'
+import Navigator from './navigator/navigator'
+import Intro from './tutorial/intro'
+import introText from './tutorial/introText'
 import {mediaRecorder, startRecording, stopRecording} from './audio/audio'
 import loadAllAssets from './assets/assets'
-import { startSyncingPhaseAndStrikes } from './firebase'
+import { startSyncingPhaseAndStrikes } from './firebase/firebase'
 import store from './store'
 import { setNavigatorStatus, setDriverStatus } from './reducers/strike-phase'
 import 'aframe-daydream-controller-component'
