+++ browser/navigator/navConsole.js
@@ -1,12 +1,12 @@
 import React from 'react'
 import Rules from './rules'
 import Panels from './navinstrucpanels'
-import PhasesWidget from './navigator/phases'
-import StrikesWidget from './navigator/strikes'
-import NavigationBar from './navigator/navBar'
-import CountDown from './navigator/clock'
-import NavContent from './navigator/navContent'
-import Footer from './navigator/footer'
+import PhasesWidget from './phases'
+import StrikesWidget from './strikes'
+import NavigationBar from './navBar'
+import CountDown from './clock'
+import NavContent from './navContent'
+import Footer from './footer'
 import PushToTalk from './pushToTalk'
 
 const SPACE_BAR = 32
