+++ browser/navigator/navContent.js
@@ -2,7 +2,7 @@ import React from 'react'
 import {Row, Navbar, NavItem} from 'react-materialize'
 import Rules from './rules'
 import Panels from './navinstrucpanels'
-import {panel1, panel2, panel3} from '../panelData'
+import {panel1, panel2, panel3} from './panelData'
 
 export default class NavigationBar extends React.Component {
   constructor(props) {
