+++ browser/navigator/navigator.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import {connect} from 'react-redux'
-import NavConsole from './container/NavConsole'
-import {setUpRecording} from '.audio/audio'
+import NavConsole from '../container/NavConsole'
+import {setUpRecording} from '../audio/audio'
 
 export default class Navigator extends React.Component {
   constructor(props) {
